Move Loggy Code To Shared Code Folder
