--- packages/ruby/loggy/lib/loggy/logger_observer.rb
@@ -1,7 +1,0 @@
-module Loggy
-    class LoggerObserver
-        def process_log_message(time_stamped_log_message)
-            puts time_stamped_log_message
-        end
-    end
-end
