--- packages/ruby/loggy/lib/loggy/observable_logger.rb
@@ -1,10 +1,0 @@
-module Loggy
-    class ObservableLogger
-        def initialize()
-            @logger_observers = []
-        end
-        def subscribe(logger_observer)
-            @logger_observers.append(logger_observer)
-        end
-    end
-end
