--- packages/ruby/loggy/lib/loggy/time_stamped_log_message.rb
@@ -1,14 +1,0 @@
-module Loggy
-    class TimeStampedLogMessage
-        def initialize(log_message)
-            @current_time_stamp = Time.now.utc.iso8601
-            @log_message = log_message
-        end
-        def time_stamp
-            return @current_time_stamp
-        end
-        def log_message
-            return @log_message
-        end
-    end
-end
