+++ packages/ruby/shared-code/loggy/logger_observer.rb
@@ -1,0 +1,5 @@
+class LoggerObserver
+    def process_log_message(time_stamped_log_message)
+        puts time_stamped_log_message
+    end
+end
