+++ packages/ruby/shared-code/loggy/observable_logger.rb
@@ -1,0 +1,8 @@
+class ObservableLogger
+    def initialize()
+        @logger_observers = []
+    end
+    def subscribe(logger_observer)
+        @logger_observers.append(logger_observer)
+    end
+end
